Adds endpoints for getting next delivery and finishing delivery
Since a GET call should be idempotent (subsequent calls shouldn't change the internal state of the application or the data), the delivery is a two stage process.
1. A GET call to /nextDelivery determines the orders for the delivery, stores them in an internal in-memory list and returns them. Those orders remain in the database, but are filtered from GET calls to /orders etc. The determination of queue position and wait time also ignores them. Subsequent GET calls to /nextDelivery simply return this list.
2. A DELETE call to /nextDelivery deletes the aforementioned list and also the orders from the database. This allows a new delivery to be prepared.
The timestamp for determining the wait time is based on the time of the first call, though there are arguments for changing it to the second one.